backend/requirements.txt
requests==2.33.1
fastapi==0.135.3
pydantic==2.13.0b3
uvicorn==0.44.0
langchain==1.2.15
langgraph==1.1.6
langchain_community==1.0.0a1
grpcio==1.80.0
grpcio-tools==1.80.0
yandexcloud==0.385.0
yandex-chain==0.0.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiomysql==0.2.0
python-dateutil==2.8.2
aiofiles==23.2.1

# pip install -r data/requirements.txt
